+++ data_structures/disjoint_sets/disjoint_sets.rb
@@ -1,4 +1,23 @@
-require "./data_structures/disjoint_sets/node.rb"
+class Node
+    attr_accessor :data, :parent, :rank
+    def initialize(data)
+        @data = data
+        @parent = self
+        @rank = 0
+    end
+    def parent
+        @parent
+    end
+    def parent=(parent)
+        @parent = parent;
+    end
+    def rank
+        @rank
+    end
+    def rank=(rank)
+        @rank = rank
+    end
+end
 
 class DisjointSets
     def make_set(d)
